--- src/ARCHIVE/MAKEFILE
@@ -1,50 +1,0 @@
-renderer.tos: renderer.o init.o clr_scrn.o clr_reg.o plt_pxl.o plt_hor.o plt_vert.o plt_line.o plt_rec.o plt_sqr.o plt_tri.o bmap_8.o bmap_16.o bmap_32.o plt_char.o plt_str.o
-	cc68x -g renderer.o init.o clr_scrn.o clr_reg.o plt_pxl.o plt_hor.o plt_vert.o plt_line.o plt_rec.o plt_sqr.o plt_tri.o bmap_8.o bmap_16.o bmap_32.o plt_char.o plt_str.o -o renderer.tos
-
-renderer.o: renderer.c raster.h
-	cc68x -g -c renderer.c
-
-init.o: init.c raster.h
-	cc68x -g -c init.c
-
-clr_scrn.o: clr_scrn.s
-	cc68x -g -c clr_scrn.s
-
-clr_reg.o: clr_reg.s
-	cc68x -g -c clr_reg.s
-
-plt_pxl.o: plt_pxl.s
-	cc68x -g -c plt_pxl.s
-
-plt_hor.o: plt_hor.s
-	cc68x -g -c plt_hor.s
-
-plt_vert.o: plt_vert.s
-	cc68x -g -c plt_vert.s
-
-plt_line.o: plt_line.s
-	cc68x -g -c plt_line.s
-
-plt_rec.o: plt_rec.s
-	cc68x -g -c plt_rec.s
-
-plt_sqr.o: plt_sqr.s
-	cc68x -g -c plt_sqr.s
-
-plt_tri.o: plt_tri.s
-	cc68x -g -c plt_tri.s
-
-bmap_8.o: bmap_8.s
-	cc68x -g -c bmap_8.s
-
-bmap_16.o: bmap_16.s
-	cc68x -g -c bmap_16.s
-
-bmap_32.o: bmap_32.s
-	cc68x -g -c bmap_32.s
-
-plt_char.o: plt_char.s
-	cc68x -g -c plt_char.s
-
-plt_str.o: plt_str.s
-	cc68x -g -c plt_str.s
